requirements.txt
fastapi==0.128.0
gradio==6.3.0
langchain==1.2.6
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_openai==1.1.7
numpy==2.4.1
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.1
